// src/app/model/customerCustomConfig.ts
export class CustomerCustomConfig{
        id: Number;
        status: Boolean;
        occOrgId: Number;
        occInternalName: String;
        occFieldType: String;
        occName: String;
        occBusinessType: String;
        occMinLength: String;
        occMaxLength: String;
        occToolTip: String;
        occMandatory: Boolean;
        occDefaultValue: Boolean;
        occDisplayInWeb: Boolean;
        occDisplayInMobile: Boolean;
        occDisplayInList: Boolean;
        occIntegrationLabel: String;
        occDisplayOrder: Number;
        occAttributeType: String;
        occDataType: String;
        occDatasource: String;
        occData: String;
        occDefault: Boolean;
        occDisplayName: String;
        occDisplayStatus: Boolean;

}